+++ pumpkin/src/plugin/api/events/player/mod.rs
@@ -1,5 +1,5 @@
-pub mod player_advancement_done;
-pub mod player_animation;
+pub mod advancement_done;
+pub mod animation;
 pub mod player_change_world;
 pub mod player_chat;
 pub mod player_command_send;
@@ -12,7 +12,7 @@ pub mod player_join;
 pub mod player_leave;
 pub mod player_login;
 pub mod player_move;
-pub mod player_pre_login;
+pub mod pre_login;
 pub mod player_teleport;
 
 use std::sync::Arc;
